[main] Start implementing command pattern.

+++ Models/ArchiveFileSource.cs
@@ -2,6 +2,6 @@ namespace mcy.Tools.Models;
 
 public class ArchiveFileSource
 {
-    public int TotalBytes{get;set;}
+    public long TotalBytes{get;set;}
     public List<ArchiveFileProperties> Files = new();
 }
